fix: don't cap messages in loadOlderMessages to preserve fetched history
Removing the capMessages call from loadOlderMessages prevents just-fetched
older messages from being immediately discarded when the merged array exceeds
MESSAGE_WINDOW_SIZE. The sliding window cap on every streaming append
(addMessage, session.message, etc.) will naturally trim during the next agent
turn. Also tightens the `before` param in message.list to z.number().int()
since sequence numbers are always integers.

+++ packages/contracts/src/ws/methods.ts
@@ -126,7 +126,7 @@ export const WS_METHODS = {
     params: z.object({
       threadId: z.string(),
       limit: z.number(),
-      before: z.number().optional(),
+      before: z.number().int().optional(),
     }),
     result: z.array(MessageSchema),
   },
